requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt --strip-extras requirements-dev.in
#
-e .
    # via -
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
build==1.4.2
    # via
    #   meeting-notes-extractor
    #   pip-tools
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
click==8.3.1
    # via
    #   pip-tools
    #   uvicorn
distro==1.9.0
    # via openai
fastapi==0.135.2
    # via meeting-notes-extractor
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   meeting-notes-extractor
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
iniconfig==2.3.0
    # via pytest
jiter==0.13.0
    # via openai
librt==0.8.1
    # via mypy
mypy==1.19.1
    # via meeting-notes-extractor
mypy-extensions==1.1.0
    # via mypy
openai==1.109.1
    # via meeting-notes-extractor
packaging==26.0
    # via
    #   build
    #   pytest
    #   wheel
pathspec==1.0.4
    # via mypy
pip-tools==7.5.3
    # via meeting-notes-extractor
pluggy==1.6.0
    # via pytest
pydantic==2.12.5
    # via
    #   fastapi
    #   openai
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via pytest
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.2
    # via meeting-notes-extractor
python-dotenv==1.2.2
    # via uvicorn
pyyaml==6.0.3
    # via uvicorn
ruff==0.15.8
    # via meeting-notes-extractor
sniffio==1.3.1
    # via openai
starlette==1.0.0
    # via fastapi
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn==0.42.0
    # via meeting-notes-extractor
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
wheel==0.46.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
